refactor: AttachmentPicker rename propTypes to AttachmentPickerPropTypes
